Split off and tidy up kodi api
